# Accumulate 16-bit float sums in 32-bit accumulators