Stack from [ghstack](https://github.com/ezyang/ghstack):
* #60389 WIP: Accumulate float sums in double accumulators
* **#60387 Accumulate 16-bit float sums in 32-bit accumulators**

Fixes gh-59489

Using 32-bit accumulators is a win-win: improved precision and improved
performance since the half precision types needed to be converted back and forth
to 32-bit float to do the arithmetic anyway.

Note that on multi-threaded or dis-contiguous sums, there can be partial sums
stored in the output so they are necessarily trucated to 16-bit. Fixing this
would require a rework of TensorIterator reductions.

Differential Revision: [D29447187](https://our.internmc.facebook.com/intern/diff/D29447187)