@peterbell10 2021-06-21
MacOS failure appears to be unrelated, probably due to a change in behaviour of the recently released SciPy 1.7 (scipy/scipy#13362).