@ngimel 2021-06-28
This looks good, can you post before and after benchmarks for fp32 (to make sure it doesn't regress) and bfloat16?