@peterbell10 2021-06-29
I've benchmarked with this simple `timeit` loop:
```python
import torch
torch.set_num_threads(1)
for dtype in [torch.float, torch.bfloat16]:
    print(dtype)
    for i in range(1, 6):
        a = torch.rand(10**i, 1000, dtype=dtype)
        print(a.shape)
        %timeit a.sum(0)
        %timeit a.sum(1)
        b = torch.rand(10**i, 2000, dtype=dtype)[:, ::2]
        %timeit b.sum(0)
        %timeit b.sum(1)
```


#### Float results
Float performance is the same, to within measurement error

|        Shape | Dim | Master Contiguous (us) | This PR Contiguous (us) | Master Discontiguous (us) | This PR Discontiguous (us) |
|-------------:|-----|:----------------------:|:-----------------------:|:-------------------------:|:--------------------------:|
|     10, 1000 | 0   |          1.92          |           1.92          |            4.33           |            4.34            |
|              | 1   |          1.71          |           1.72          |            2.86           |            2.93            |
|    100, 1000 | 0   |          6.42          |           6.84          |            27.7           |            27.3            |
|              | 1   |          5.65          |           6.01          |            17.2           |            17.1            |
|   1000, 1000 | 0   |           57           |           56.2          |            259            |             256            |
|              | 1   |          41.7          |           41.5          |            150            |             152            |
|  10000, 1000 | 0   |          1,360         |          1,340          |           8,020           |            7,890           |
|              | 1   |           838          |           823           |           3,030           |            2,970           |
| 100000, 1000 | 0   |         24,900         |          24,200         |           92,200          |           91,400           |
|              | 1   |         12,100         |          11,800         |           30,900          |           31,100           |

#### BFloat16 results
For BFloat16 this is consistently around 3x faster

|        Shape | Dim | Master Contiguous (us) | This PR Contiguous (us) | Master Discontiguous (us) | This PR Discontiguous (us) |
|-------------:|-----|:----------------------:|:-----------------------:|:-------------------------:|:--------------------------:|
|     10, 1000 | 0   |          5.19          |           3.51          |            17.6           |            7.29            |
|              | 1   |          4.15          |           2.00          |            13.4           |            4.39            |
|    100, 1000 | 0   |          22.7          |           7.02          |            124            |            40.9            |
|              | 1   |          29.1          |           8.02          |            122            |            31.6            |
|   1000, 1000 | 0   |           213          |           50.1          |           1,190           |             404            |
|              | 1   |           283          |           68.2          |           1,210           |             303            |
|  10000, 1000 | 0   |          3,760         |           921           |           14,500          |            9,230           |
|              | 1   |          2,860         |           679           |           12,200          |            3,110           |
| 100000, 1000 | 0   |         82,500         |          20,700         |          151,000          |           99,100           |
|              | 1   |         29,400         |          8,590          |          122,000          |           31,900           |