@ngimel 2021-06-29
Internal builds are failing with 
```
caffe2/aten/src/ATen/native/cpu/SumKernel.cpp:351:17: error: unused variable 'vec_size' [-Werror,-Wunused-variable]
  const int64_t vec_size = size0 / vacc_t::size();
```